backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.28
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.6.3
pydantic-settings==2.2.1
redis==5.0.2
httpx==0.27.0
apscheduler==3.10.4
pycryptodome==3.20.0
pyyaml==6.0.1
python-multipart==0.0.9
